Revert "chore: Rename files to match annotations"
This reverts commit e6caae2ceea347f4c94f4036fb165261072784e5.
